Give details in cloneWithType errors
