+++ langlib/lang.value/src/main/java/org/ballerinalang/langlib/value/CloneWithType.java
@@ -56,8 +56,9 @@ import java.util.Map;
 import java.util.Set;
 
 import static io.ballerina.runtime.api.creators.ErrorCreator.createError;
-import static io.ballerina.runtime.internal.util.exceptions.BallerinaErrorReasons.CONSTRUCT_FROM_CONVERSION_ERROR;
-import static io.ballerina.runtime.internal.util.exceptions.BallerinaErrorReasons.CONSTRUCT_FROM_CYCLIC_VALUE_REFERENCE_ERROR;
+import static io.ballerina.runtime.internal.ErrorUtils.createConversionError;
+import static io.ballerina.runtime.internal.util.exceptions.BallerinaErrorReasons.VALUE_LANG_LIB_CONVERSION_ERROR;
+import static io.ballerina.runtime.internal.util.exceptions.BallerinaErrorReasons.VALUE_LANG_LIB_CYCLIC_VALUE_REFERENCE_ERROR;
 import static io.ballerina.runtime.internal.util.exceptions.RuntimeErrors.INCOMPATIBLE_CONVERT_OPERATION;
 
 /**
@@ -89,7 +90,7 @@ public class CloneWithType {
         } catch (BError e) {
             return e;
         } catch (BallerinaException e) {
-            return createError(CONSTRUCT_FROM_CONVERSION_ERROR, StringUtils.fromString(e.getDetail()));
+            return createError(VALUE_LANG_LIB_CONVERSION_ERROR, StringUtils.fromString(e.getDetail()));
         }
     }
 
@@ -105,16 +106,30 @@ public class CloneWithType {
             if (targetType.isNilable()) {
                 return null;
             }
-            return createError(CONSTRUCT_FROM_CONVERSION_ERROR,
+            return createError(VALUE_LANG_LIB_CONVERSION_ERROR,
                     BLangExceptionHelper.getErrorMessage(RuntimeErrors.CANNOT_CONVERT_NIL, targetType));
         }
+
+        List<String> errors = new ArrayList<>();
         Set<Type> convertibleTypes;
-        convertibleTypes = TypeConverter.getConvertibleTypes(value, targetType, null, false,
-                new ArrayList<>());
+        convertibleTypes = TypeConverter.getConvertibleTypes(value, targetType, null, false, errors);
 
         Type sourceType = TypeChecker.getType(value);
         if (convertibleTypes.isEmpty()) {
-            throw createConversionError(value, targetType);
+            if (errors.isEmpty()) {
+                throw createConversionError(value, targetType);
+            } else {
+                StringBuilder errorMsg = new StringBuilder();
+                byte errorCount = 0;
+                for (String error : errors) {
+                    errorMsg.append("\n\t\t").append(error);
+                    errorCount++;
+                    if (errorCount == 100) {
+                        break;
+                    }
+                }
+                throw createConversionError(value, targetType, errorMsg.toString());
+            }
         } else if (!allowAmbiguity && convertibleTypes.size() > 1 && !convertibleTypes.contains(sourceType) &&
                 !TypeConverter.hasIntegerSubTypes(convertibleTypes)) {
             throw createAmbiguousConversionError(value, targetType);
@@ -144,7 +159,7 @@ public class CloneWithType {
         TypeValuePair typeValuePair = new TypeValuePair(value, targetType);
 
         if (unresolvedValues.contains(typeValuePair)) {
-            throw new BallerinaException(CONSTRUCT_FROM_CYCLIC_VALUE_REFERENCE_ERROR.getValue(),
+            throw new BallerinaException(VALUE_LANG_LIB_CYCLIC_VALUE_REFERENCE_ERROR.getValue(),
                                          BLangExceptionHelper
                                                  .getErrorMessage(RuntimeErrors.CYCLIC_VALUE_REFERENCE, value.getType())
                                                  .getValue());
@@ -175,7 +190,7 @@ public class CloneWithType {
                 break;
             default:
                 // should never reach here
-                throw CloneUtils.createConversionError(value, targetType);
+                throw createConversionError(value, targetType);
         }
 
         unresolvedValues.remove(typeValuePair);
@@ -219,7 +234,7 @@ public class CloneWithType {
                 break;
         }
         // should never reach here
-        throw CloneUtils.createConversionError(map, targetType);
+        throw createConversionError(map, targetType);
     }
 
     private static BMap<BString, Object> convertToRecord(BMap<?, ?> map, List<TypeValuePair> unresolvedValues,
@@ -291,7 +306,7 @@ public class CloneWithType {
                 break;
         }
         // should never reach here
-        throw CloneUtils.createConversionError(array, targetType);
+        throw createConversionError(array, targetType);
     }
 
     private static Object convertTable(BTable<?, ?> bTable, Type targetType,
@@ -314,17 +329,10 @@ public class CloneWithType {
         return ValueCreator.createTableValue(tableType, data, fieldNames);
     }
 
-    private static BError createConversionError(Object inputValue, Type targetType) {
-        return createError(CONSTRUCT_FROM_CONVERSION_ERROR,
-                           BLangExceptionHelper.getErrorMessage(INCOMPATIBLE_CONVERT_OPERATION,
-                        TypeChecker.getType(inputValue), targetType));
-    }
-
     private static BError createAmbiguousConversionError(Object inputValue, Type targetType) {
-        return createError(CONSTRUCT_FROM_CONVERSION_ERROR,
+        return createError(VALUE_LANG_LIB_CONVERSION_ERROR,
                            BLangExceptionHelper.getErrorMessage(INCOMPATIBLE_CONVERT_OPERATION,
                                                                 TypeChecker.getType(inputValue), targetType)
                                    .concat(StringUtils.fromString(": ".concat(CloneWithType.AMBIGUOUS_TARGET))));
     }
-
 }
